requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
python-dateutil==2.8.2
